requirements.txt
# 基础依赖（同步自 pyproject.toml [project.dependencies]）
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
alembic>=1.13
pydantic-settings>=2.0
passlib[bcrypt]>=1.7
bcrypt<4.1
pypinyin>=0.52
python-multipart>=0.0.9
itsdangerous>=2.2

# PostgreSQL 支持（同步自 pyproject.toml [project.optional-dependencies.pg]）
psycopg[binary]>=3.0
